requirement.txt
-e ./fairseq/

absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.8
anyio==3.7.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bitarray==2.7.3
brotlipy==0.7.0
cachetools==5.3.0
-e git+https://github.com/ruotianluo/self-critical.pytorch.git@be1a526b058707eb136ede9397b54d971374fb1f#egg=captioning
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
conda==4.12.0
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.11.0
Cython==0.29.34
datasets==2.12.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
dill==0.3.6
Distance==0.1.3
dnspython==2.1.0
editdistance==0.6.2
einops==0.6.1
exceptiongroup==1.1.1
fastapi==0.96.0
ffmpy==0.3.0
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.0.3
g2p-en==2.1.0
gensim==4.3.1
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gradio==3.34.0
gradio-client==0.2.6
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
hydra-core==1.0.7
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==6.0.4
ipython @ file:///tmp/build/80754af9/ipython_1614615245721/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-loader==0.2
Levenshtein==0.21.0
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
librosa==0.10.0.post2
lightning-utilities==0.8.0
linkify-it-py==2.0.2
llvmlite==0.40.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
omegaconf==2.0.6
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.9.0
packaging==23.1
pandas==2.0.1
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pkginfo==1.7.0
pkgutil-resolve-name==1.3.10
pooch==1.6.0
portalocker==2.7.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.20.0rc1
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.4
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.10.7
pydicom==2.3.1
pydub==0.25.1
Pygments==2.15.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.0.9
pypinyin==0.47.1
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
python-Levenshtein==0.21.0
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.4.1
PyYAML==5.4.1
rapidfuzz==3.0.0
regex==2023.5.5
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.1.2
rsa==4.9
ruamel-yaml==0.15.87
sacrebleu==2.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
semantic-version==2.10.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
smart-open==6.3.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve @ file:///tmp/build/80754af9/soupsieve_1614023247537/work
soxr==0.3.5
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.4.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.13
toolz==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchelastic==0.2.2
torchmetrics==0.11.4
torchtext==0.12.0
torchvision==0.12.0
tqdm==4.62.1
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uvicorn==0.22.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
websockets==11.0.3
Werkzeug==2.3.3
xxhash==3.2.0
yarl==1.9.2
zhconv==1.4.3
zipp==3.15.0
